requirements.txt
apache-libcloud
cryptography
requests
